refactor: split the run orchestrator god-function into phase modules
runAll was a ~275-line function (depth-10 nesting, a labeled break) fusing five
jobs. Extract the two big inline blocks so runAll becomes a thin orchestrator —
resolve -> preflight -> execute -> report — dropping 843 -> 278 lines:

- baselineScanner.ts: the MCP pre-flight scans as a Chain of Responsibility
  (runBaselineScans + the resource/tool-description/rug-pull scanners). Each
  scanner contributes an EvaluatorResult or nothing; runAll pushes them in order.
- evaluatorLoop.ts: runEvaluatorAttacks — the evaluator attack loop. The labeled
  `break evaluatorLoop` becomes a return; captureSessionContext moves in. The two
  verbatim-duplicated stop-error branches (isStopError vs TargetStopError) and the
  thrice-repeated evaluatorMeta object are consolidated (DRY), and makeFailedResult
  now uses errorJudge().

Backward-compatible: runAll(config, options) keeps its exact signature, so all
callers (CLI, SDK, MCP, browser) are untouched. The RunEngine/RunRequest boundary
from the plan is intentionally deferred — it would rewire the front-ends and
change the public API. Behavior-preserving: evaluatorResults ordering (scans then
evaluators), partial-report-on-stop, mcpTarget.close() in a finally, topo-sort,
dependency-skip, and upstream-session threading are all reproduced; the
runAll.smoke and orchestrator.equivalence drift guards stay green.

New baselineScanner test covers the chain (the MCP pre-flight had no coverage).
