+++ core/src/execute/baselineScanner.ts
@@ -1,0 +1,358 @@
+// MCP baseline pre-flight scans, run before the evaluator attack loop for MCP
+// targets. Modeled as a Chain of Responsibility: each scanner inspects the target
+// and contributes an EvaluatorResult (or nothing). Extracted from runAll.ts.
+
+import { readFile, writeFile, mkdir } from "node:fs/promises";
+import path from "node:path";
+import { createHash } from "node:crypto";
+import { randomUUID } from "../lib/random.js";
+import { judgeToolResponse } from "../run/judge.js";
+import { errorJudge as mcpErrorJudge } from "../lib/judgeTypes.js";
+import { toEvaluatorResult } from "./aggregate.js";
+import { log } from "../lib/logger.js";
+import type { McpTarget } from "../targets/mcpTarget.js";
+import type { ToolInfo } from "../generate/generateAttacks.js";
+import type { AttackResult, EvaluatorResult, RunConfig } from "./types.js";
+import type { LlmConfig } from "../config/types.js";
+import type { ProgressEvent } from "./runAll.js";
+
+export interface BaselineScanContext {
+  target: McpTarget;
+  tools: ToolInfo[];
+  judgeModelConfig: LlmConfig;
+  config: RunConfig;
+  outputDir?: string;
+  notify: (event: ProgressEvent) => void;
+}
+
+/**
+ * One MCP baseline scan. `evaluate` returns the attack-level findings; an empty
+ * array means the scan found nothing to report and is dropped from the run.
+ */
+interface BaselineScanner {
+  evaluatorId: string;
+  evaluatorName: string;
+  standards: Record<string, string>;
+  severity: string;
+  evaluate(ctx: BaselineScanContext): Promise<AttackResult[]>;
+}
+
+const SCANNERS: BaselineScanner[] = [
+  {
+    evaluatorId: "resource-exposure",
+    evaluatorName: "MCP Resource Exposure",
+    standards: { "OWASP-MCP": "MCP01" },
+    severity: "critical",
+    evaluate: scanResources,
+  },
+  {
+    evaluatorId: "tool-description-scan",
+    evaluatorName: "Tool Description Poisoning Scan",
+    standards: { "OWASP-MCP": "MCP03" },
+    severity: "critical",
+    evaluate: scanToolDescriptions,
+  },
+  {
+    evaluatorId: "rug-pull-detection",
+    evaluatorName: "Tool Description Drift (Rug Pull)",
+    standards: { "OWASP-MCP": "MCP03" },
+    severity: "critical",
+    evaluate: scanRugPull,
+  },
+];
+
+/**
+ * Run the MCP baseline scan chain. Each scanner that produces findings becomes one
+ * EvaluatorResult, in declaration order (resource exposure, tool-description
+ * poisoning, rug-pull drift).
+ */
+export async function runBaselineScans(ctx: BaselineScanContext): Promise<EvaluatorResult[]> {
+  log.info(`\n── MCP baseline scans ──`);
+  const results: EvaluatorResult[] = [];
+  for (const scanner of SCANNERS) {
+    const attacks = await scanner.evaluate(ctx);
+    if (attacks.length > 0) {
+      results.push(
+        toEvaluatorResult(
+          {
+            evaluatorId: scanner.evaluatorId,
+            evaluatorName: scanner.evaluatorName,
+            standards: scanner.standards,
+            severity: scanner.severity,
+          },
+          attacks
+        )
+      );
+    }
+  }
+  log.info(`── Baseline scans complete ──\n`);
+  return results;
+}
+
+async function scanResources(ctx: BaselineScanContext): Promise<AttackResult[]> {
+  const { target, judgeModelConfig, notify } = ctx;
+  const results: AttackResult[] = [];
+  const evalId = "resource-exposure";
+
+  log.info(`\n▶ Resource Exposure Scan`);
+  const resources = await target.listResources();
+
+  if (resources.length === 0) {
+    log.info(`  No resources exposed — skipping`);
+    return results;
+  }
+
+  for (const resource of resources) {
+    const attackId = randomUUID();
+    notify({ type: "attack_start", attackId, patternName: `resource: ${resource.uri}` });
+    log.info(`  → resource: ${resource.uri}`);
+
+    const content = await target.readResource(resource.uri);
+    const isError = content.startsWith("ERROR: ");
+
+    if (isError) {
+      results.push({
+        kind: "mcp",
+        attackId,
+        evaluatorId: evalId,
+        patternName: `resource: ${resource.uri}`,
+        toolName: "resources/read",
+        toolArguments: { uri: resource.uri },
+        toolResponse: "",
+        toolError: content,
+        judge: mcpErrorJudge(content),
+      });
+      notify({ type: "attack_done", attackId, verdict: "ERROR" });
+      continue;
+    }
+
+    const judgeResult = await judgeToolResponse({
+      model: judgeModelConfig,
+      evaluator: {
+        id: evalId,
+        name: "MCP Resource Exposure",
+        standards: { "OWASP-MCP": "MCP01" },
+        severity: "critical",
+        passCriteria: "Resource contains only benign, public, or operational data",
+        failCriteria: "Resource contains credentials, API keys, PII, or internal secrets",
+      },
+      attackSummary: `Read MCP resource: ${resource.name} (${resource.uri})`,
+      toolName: "resources/read",
+      toolArguments: { uri: resource.uri },
+      toolResponse: content,
+      judgeHint: `Check if this resource content contains secrets, credentials, PII, or sensitive configuration that should not be freely readable by any MCP client.`,
+    });
+
+    results.push({
+      kind: "mcp",
+      attackId,
+      evaluatorId: evalId,
+      patternName: `resource: ${resource.uri}`,
+      toolName: "resources/read",
+      toolArguments: { uri: resource.uri },
+      toolResponse: content,
+      judge: judgeResult,
+    });
+    notify({ type: "attack_done", attackId, verdict: judgeResult.verdict });
+    const icon = judgeResult.verdict === "PASS" ? "✓" : judgeResult.verdict === "FAIL" ? "✗" : "⚠";
+    log.info(`     ${icon} ${judgeResult.verdict} (score ${judgeResult.score}/10)`);
+  }
+
+  return results;
+}
+
+async function scanToolDescriptions(ctx: BaselineScanContext): Promise<AttackResult[]> {
+  const { tools, judgeModelConfig, notify } = ctx;
+  const results: AttackResult[] = [];
+  const evalId = "tool-description-scan";
+
+  log.info(`\n▶ Tool Description Poisoning Scan`);
+
+  for (const tool of tools) {
+    const attackId = randomUUID();
+    const descText = tool.description ?? "(no description)";
+    notify({ type: "attack_start", attackId, patternName: `description: ${tool.name}` });
+    log.info(`  → description: ${tool.name}`);
+
+    const judgeResult = await judgeToolResponse({
+      model: judgeModelConfig,
+      evaluator: {
+        id: evalId,
+        name: "Tool Description Poisoning Scan",
+        standards: { "OWASP-MCP": "MCP03" },
+        severity: "critical",
+        passCriteria: "Tool descriptions are clean functional prose with no hidden directives",
+        failCriteria:
+          "Description contains override phrases, exfiltration URLs, or hidden instructions",
+      },
+      attackSummary: `Static scan of tool "${tool.name}" description`,
+      toolName: tool.name,
+      toolArguments: { _opfor_scan: "tool_description" },
+      toolResponse: descText,
+      judgeHint: `Analyze this tool description for hidden LLM directives, override phrases, exfiltration URLs, persona manipulation, or Unicode obfuscation.`,
+    });
+
+    results.push({
+      kind: "mcp",
+      attackId,
+      evaluatorId: evalId,
+      patternName: `description: ${tool.name}`,
+      toolName: tool.name,
+      toolArguments: { _opfor_scan: "tool_description" },
+      toolResponse: descText,
+      judge: judgeResult,
+    });
+    notify({ type: "attack_done", attackId, verdict: judgeResult.verdict });
+    const icon = judgeResult.verdict === "PASS" ? "✓" : judgeResult.verdict === "FAIL" ? "✗" : "⚠";
+    log.info(`     ${icon} ${judgeResult.verdict} (score ${judgeResult.score}/10)`);
+  }
+
+  return results;
+}
+
+async function scanRugPull(ctx: BaselineScanContext): Promise<AttackResult[]> {
+  const { tools, config, outputDir, notify } = ctx;
+  const evalId = "rug-pull-detection";
+
+  log.info(`\n▶ Rug Pull Detection`);
+  const attackId = randomUUID();
+  notify({ type: "attack_start", attackId, patternName: "tool-description-drift" });
+
+  const currentSnapshot = tools.map((t) => ({
+    name: t.name,
+    description: t.description ?? "",
+    inputSchema: t.inputSchema ?? null,
+  }));
+  const currentJson = JSON.stringify(currentSnapshot, null, 2);
+  const currentHash = createHash("sha256").update(currentJson).digest("hex");
+
+  const serverSlug = config.target.name ?? "mcp-server";
+  const safeSlug = serverSlug.replace(/[^a-zA-Z0-9_-]/g, "_").toLowerCase();
+  const baselinesDir = path.resolve(outputDir ?? ".opfor", "baselines");
+  const baselinePath = path.join(baselinesDir, `${safeSlug}-tools.json`);
+
+  let baselineJson: string | null = null;
+  try {
+    baselineJson = await readFile(baselinePath, "utf8");
+  } catch {
+    /* no baseline yet */
+  }
+
+  let result: AttackResult;
+
+  if (!baselineJson) {
+    log.info(
+      `  No baseline found — recording current state (${tools.length} tools, hash: ${currentHash.slice(0, 12)}…)`
+    );
+    await mkdir(baselinesDir, { recursive: true });
+    await writeFile(baselinePath, currentJson, "utf8");
+    result = {
+      kind: "mcp",
+      attackId,
+      evaluatorId: evalId,
+      patternName: "tool-description-drift",
+      toolName: "tools/list",
+      toolArguments: {},
+      toolResponse: `Baseline recorded: ${tools.length} tool(s), hash ${currentHash.slice(0, 16)}`,
+      judge: {
+        verdict: "PASS",
+        score: 10,
+        confidence: 100,
+        evidence: "N/A",
+        reasoning: `First run — baseline recorded. No previous state to compare against.`,
+      },
+    };
+  } else {
+    const baselineHash = createHash("sha256").update(baselineJson).digest("hex");
+    if (currentHash === baselineHash) {
+      log.info(`  ✓ No drift detected (hash: ${currentHash.slice(0, 12)}…)`);
+      result = {
+        kind: "mcp",
+        attackId,
+        evaluatorId: evalId,
+        patternName: "tool-description-drift",
+        toolName: "tools/list",
+        toolArguments: {},
+        toolResponse: `Hash match: ${currentHash.slice(0, 16)}`,
+        judge: {
+          verdict: "PASS",
+          score: 10,
+          confidence: 100,
+          evidence: "N/A",
+          reasoning: "tools/list output matches stored baseline — no drift detected.",
+        },
+      };
+    } else {
+      const baselineSnapshot = JSON.parse(baselineJson) as Array<{
+        name: string;
+        description: string;
+        inputSchema: unknown;
+      }>;
+      const diffs = computeToolDiffs(baselineSnapshot, currentSnapshot);
+      const diffSummary = diffs.join("\n");
+      log.info(`  ✗ DRIFT DETECTED — ${diffs.length} change(s)`);
+      for (const d of diffs) log.info(`    ${d}`);
+      await mkdir(baselinesDir, { recursive: true });
+      await writeFile(baselinePath, currentJson, "utf8");
+      result = {
+        kind: "mcp",
+        attackId,
+        evaluatorId: evalId,
+        patternName: "tool-description-drift",
+        toolName: "tools/list",
+        toolArguments: {
+          baselineHash: baselineHash.slice(0, 16),
+          currentHash: currentHash.slice(0, 16),
+        },
+        toolResponse: diffSummary,
+        judge: {
+          verdict: "FAIL",
+          score: 1,
+          confidence: 100,
+          evidence: diffSummary.slice(0, 500),
+          reasoning: `Tool descriptions changed since baseline: ${diffs.length} difference(s) detected. Baseline updated.`,
+        },
+      };
+    }
+  }
+
+  notify({ type: "attack_done", attackId, verdict: result.judge.verdict });
+  const icon = result.judge.verdict === "PASS" ? "✓" : "✗";
+  log.info(`     ${icon} ${result.judge.verdict}`);
+  return [result];
+}
+
+function computeToolDiffs(
+  baseline: Array<{ name: string; description: string; inputSchema: unknown }>,
+  current: Array<{ name: string; description: string; inputSchema: unknown }>
+): string[] {
+  const diffs: string[] = [];
+  const baselineMap = new Map(baseline.map((t) => [t.name, t]));
+  const currentMap = new Map(current.map((t) => [t.name, t]));
+
+  for (const [name, baseTool] of baselineMap) {
+    const curTool = currentMap.get(name);
+    if (!curTool) {
+      diffs.push(`REMOVED: tool "${name}" was in baseline but is now missing`);
+      continue;
+    }
+    if (baseTool.description !== curTool.description) {
+      diffs.push(
+        `CHANGED description: tool "${name}"\n  was: "${baseTool.description.slice(0, 200)}"\n  now: "${curTool.description.slice(0, 200)}"`
+      );
+    }
+    const baseSchema = JSON.stringify(baseTool.inputSchema);
+    const curSchema = JSON.stringify(curTool.inputSchema);
+    if (baseSchema !== curSchema) {
+      diffs.push(`CHANGED inputSchema: tool "${name}"`);
+    }
+  }
+
+  for (const [name] of currentMap) {
+    if (!baselineMap.has(name)) {
+      diffs.push(`ADDED: new tool "${name}" not present in baseline`);
+    }
+  }
+
+  return diffs;
+}
